refactor: timestampToDateString を backfillDisplayFileName から抽出 (Closes #332)
PR #330 review-pr code-reuse Important #2 対応。

## 問題
- \`functions/src/pdf/pdfOperations.ts\` が本番パスで \`backfillDisplayFileName\` 経由で
  \`timestampToDateString\` を import していた
- 「backfill 専用」という scope を示唆する name と、本番で使う実態が不一致
- 将来の reader が「これは backfill でしか使われていない」と誤読するリスク

## 変更
- \`functions/src/utils/timestampHelpers.ts\` を新設し、\`TimestampLike\` interface +
  \`timestampToDateString\` を移動
- \`backfillDisplayFileName.ts\`: \`timestampHelpers\` から import、自前定義削除
- \`pdfOperations.ts\`: import path を \`../utils/timestampHelpers\` に変更
- テスト: \`timestampHelpers.test.ts\` を新設（5 cases 移動）、
  \`backfillDisplayFileName.test.ts\` は \`buildDisplayFileNameFromDoc\` のみ残す

## Phase B of WBS
Phase A (#339 #340) に続く WBS Phase B。以降 Phase C (#331 #333 #335 sanitize 整理) へ。

テスト 635 passing (loadMasterData 7 + timestampToDateString 5 + その他)、
lint 0 errors、tsc (functions/frontend) PASS、回帰なし。
